Fix issue #15, add saving bank option
This commit adds a new field to the group model, which is a boolean
field. This field is used to determine whether the group is a saving
bank or not. If the group is a saving bank, then the group will be
displayed in the nav bar. If not, then the group will not be displayed
in the nav bar.

This commit also adds a new migration file, which is used to add the
new field to the group model.